requirements.txt
Flask==2.3.3
azure-ai-formrecognizer==3.3.0
openai==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
Pillow==10.0.1
httpx>=0.25.0,<0.29.0
